backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
python-multipart
pytesseract
Pillow
pdfplumber
python-dotenv
